Parameters for Generalized Lotka-Volterra system

cNODE.getParameters — Function

getParameters(
    N::Int64,
    C::Float64,
    σ::Float64
    )

Get matrix of interactions A and instrisic growth rates r.

Arguments:

N::Int64 : number of species

C::Float64 : connectivity

σ::Float64 : interaction strength

cNODE.getGLV — Function

getGLV(
    x,
    pars::Array,
    t
)

Get GLV dynamics.

Arguments:

pars::Array : concatenation of matrix A and vector r

x and t : placeholders for ODE solver

cNODE.getIC — Function

getIC(
    N::Int64,
    M::Int64
    )

Get collection of species z and numerical initial condition x₀.

Arguments:

N::Int64 : number of species

M::Int64 : number of communities

Obtain steady states from GLV

cNODE.getSteadyState — Function

getSteadyState(
    x₀::Vector,
    τ::Float64,
    A::Array{Float64},
    r::Array{Float64}
    )

Get steady-state of dynamics by integrating ODE.

Arguments:

x₀::Vector : numerical initial condition

τ::Float64 : integration time

A::Array{Float64} : interaction matrix

r::Array{Float64} : intrinsic growth rates

Rewire network

cNODE.getRewiredNetwork — Function

getRewiredNetwork(
    A::Array{Float64},
    p::Float64
    )

Get rewired interaction matrix.

Arguments:

A::Array{Float64} : interaction matrix

p::Float64 : rewiring probability

Generate synthetic data

cNODE.generate_data — Function

    generate_data(
        N::Int64,
        M::Int64,
        repetitions::Int64,
        values::Int64,
        params::Array
    )

Generate synthetic data for parameters in params. The function generates a number of repetitions for every value of values.

Arguments:

N::Int64 : number of species

M::Int64 : number of communitites

repetitions::Int64 : number of ecological networks

values::Int64 : number of values per parameter

params::Array : sweeping parameters